Apply dirmode to top-level cache directory, too
Due to the cache directory already created in the constructor, if it doesn't
exist already it will be created with wrong permissions. The fluent setter for
dirMode is only applied after the directory was created.

Setting the directoryMode upfront via a new constructor param fixes
this, though it also obsoletes the `withDirectoryMode()` method.

Fixes #46